+++ dev_project/project_env/services/__init__.py
@@ -1,5 +1,11 @@
 from .ci_image_build import CiImageBuildService
+from .docker_base_image import BaseImageService
 from .platform_sources import PlatformSourcesService
 from .vscode_configurator import VscodeConfigurator
 
-__all__ = ["CiImageBuildService", "PlatformSourcesService", "VscodeConfigurator"]
+__all__ = [
+    "BaseImageService",
+    "CiImageBuildService",
+    "PlatformSourcesService",
+    "VscodeConfigurator",
+]